feat: new method to get parsed AST
